[LI-HOTFIX] Do not expire quota sensors
TICKET = LIKAFKA-18966
LI_DESCRIPTION =
The Kafka sensor for throttle time is recorded only when there is throttling and thus if a client is not throttled for a while, the sensor expires. And when it expires, it is not registered as an attribute of the corresponding mbean sensor. However, healthcheck expects this attribute to exist in the mbean sensor all the time and hence raising the attribute not found exception and polluting application logs.
EXIT_CRITERIA = MANUAL [""]