[LI-HOTFIX] Clearing the ZkReplicaStateMachine request batch state upon ControllerMovedException (#117)
TICKET = KAFKA-12315
LI_DESCRIPTION = Clearing the ZkReplicaStateMachine request batch state upon
ControllerMovedException so that when a node becomes the controller again, it
has a clean state and can operate normally. More details can be found at
KAFKA-12315.

EXIT_CRITERIA = When KAFKA-12315 is resolved and the commit is pulled in from upstream