[LI-HOTFIX] Introducing replica.request.timeout.ms (#234)
TICKET = LIKAFKA-40758
LI_DESCRIPTION = Currently, processing the LeaderAndIsr request (wrapped as part of the
LiCombinedControlRequest) requires blocking on network I/O, which can cause a max delay of
request.timeout.ms.
The same request.timeout.ms config is being used in 3 places within a kafka cluster
1. It's used by the controller when sending out control requests to brokers
2. It's used by a broker when sending ControlledShutdown request to the controller
3. It's used by a follower broker when sending a ListOffset or Fetch request to a leader broker.

While it makes sense to use a relatively long request.timeout.ms for the first two use cases,
using a shorter value for the 3rd use case speeds up the processing of the LeaderAndIsr request,
if the leader broker has crashed or lost its network connectivity with the follower.

This PR introduces a new config replica.request.timeout.ms to separate out the 3rd use case.

EXIT_CRITERIA = If and when this change is accepted upstream and pulled into this repo.