[LI-HOTFIX] Fix build for Java 11+ (#240)
The javax.xml.bind package was deprecated since Java 9, and has been removed from Java 11. The build.gradle script uses a call to DataTypeConverter that relies on the `javax.xml.bind` package, and hence, fails to build on any JDK at or above JDK 11. This is provided by the javax.xml.bind:jaxb-api artifact, and is already a part of Kafka's dependencies.