[LI-HOTFIX] Use max broker epoch in controller requests (#247)
TICKET = KAFKA-7186
LI_DESCERIPTION = Currently the controller is creating a separate copy of UpdateMetadataRequest for
each broker in the cluster, which may cause OOMs upon controller startup. This patch is the first step
toward making the UpdateMetadataRequests cacheable. We change the broker-specific epoch in control
requests to a common maximum broker epoch seen in all brokers, such that all UpdateMetadataRequests have
the same payload.

EXIT_CRITERIA = KAFKA-7186

* Fixed unit tests for clients module.

* Suppress NPath complexity error from Checkstyle.

* Fix core unit tests.

Co-authored-by: Lucas Wang <luwang@linkedin.com>

====This also squashes another commit====

[LI-FIXUP] Fixed the API version check when building LeaderAndIsrResponse in ReplicaManager (#262)

In PR #247, we changed the response versions for LeaderAndIsrResponse.
Version 5 became version 6. We missed changing this check that actually
transforms the response and groups it by TopicId if the version >= 5
(which should now become version 6).

Also moved the catch-all clause for the case inside KafkaApis to
maintain consistency.

https://github.com/linkedin/kafka/pull/247/files#diff-e7cce238d48a1918a4c35d38ccaf1ea6821003f3e0c85a4cd9af6e804f9ef024