[LI-HOTFIX] Add metric to show the MetadataOutgoingBytesPerSec (#328)
TICKET = KAFKA-13797
LI_DESCRIPTION =
In the past, we found that some brokers in the venice cluster went through heavy load
due to excessive Metadata requests. This PR adds a metric to show the rate of metadata outgoing
bytes per second for easier troubleshooting of such issues.

EXIT_CRITERIA = When KAFKA-13797 is resolved and the changes are pulled into this repo.