[LI-FIXUP] Revert "Improve the efficiency of replica deletion (#331)"
This reverts commit 7d784cefa37e9b0a32533c90a741348a479c1c99.

During upstream review of this change, we find that there is a
risk where the leader epoch cache is not re-initialized after replica deletion,
and then it is subsequently used by another request.
The broker could potentially crash under such sequence of events.
Thus, we are reverting this change for now until we have better understanding
of the implications.