[LI-HOTFIX] Set correct delete flag when merging control request (#340)
TICKET =
LI_DESCRIPTION =
1. When IBP >= KAFKA_2_6_IV0 and stopReplicaRequestVersion >= 4, the StopReplicaRequest.Builder's
deletePartitions flag will always be set to a placeholder value of false, and the source of truth
delete flag lies in each StopReplicaPartitionState.
This PR fixes the bug by using the source of true flag from each StopReplicaPartitionState.

2. I'm also changing the default server config in integration tests by enbaling the
LiCombinedControlRequestEnableProp, given that is the setting used by our clusters.

3. Fixing version issues when converting between the LiCombinedControl request and the original
control request types. Make sure that the error codes from a LeaderAndIsrResponse can be properly
carried back.

EXIT_CRITERIA = If/when the LiCombinedControlRequest feature is merged in upstream.